+++ script.module.codequick/lib/codequick/script.py
@@ -183,8 +183,16 @@ class Script(object):
         :param func: Function that will be called after "xbmcplugin.endOfDirectory" is called.
         :param args: "Positional" arguments that will be passed to function.
         :param kwargs: "Keyword" arguments that will be passed to function.
+
+        .. note::
+
+            There is one optional keyword only argument ``function_type``. Values are as follows.
+            * ``0`` Only run if no errors are raised. (Default)
+            * ``1`` Only run if an error has occurred.
+            * ``2`` Run regardless if an error was raised or not.
         """
-        dispatcher.register_delayed(func, args, kwargs)
+        function_type = kwargs.get("function_type", 0)
+        dispatcher.register_delayed(func, args, kwargs, function_type)
 
     @staticmethod
     def log(msg, args=None, lvl=10):